refactor(filid): split heavy-offender files into single-fn files
Split three files with 7+ functions into per-function files, preserving
all public import paths via facade re-exports.

- caches/session-cache.ts (8 fn) → 8 sibling files in caches/, original
  becomes facade re-exporting them.
- mcp/tools/debt-manage/debt-manage.ts (11 fn) → handlers/ organ with 4
  action handlers, utils/ organ with 6 private helpers; debt-manage.ts
  retains only the dispatcher and DebtManageInput type.
- lib/logger.ts (7 fn) → 9 per-function files plus log-dir-state.ts
  wrapper object (required because ES modules forbid mutating named
  imports); logger.ts becomes a facade.
