+++ packages/filid/src/core/infra/cache-manager/caches/cwd-hash.ts
@@ -1,0 +1,5 @@
+import { createHash } from 'node:crypto';
+
+export function cwdHash(cwd: string): string {
+  return createHash('sha256').update(cwd).digest('hex').slice(0, 16);
+}
