+++ packages/filid/src/core/infra/cache-manager/caches/session-id-hash.ts
@@ -1,0 +1,5 @@
+import { createHash } from 'node:crypto';
+
+export function sessionIdHash(sessionId: string): string {
+  return createHash('sha256').update(sessionId).digest('hex').slice(0, 16);
+}
