+++ packages/filid/src/lib/get-log-dir.ts
@@ -1,0 +1,6 @@
+import { logDirState } from './log-dir-state.js';
+
+/** Get current log directory (for testing). */
+export function getLogDir(): string | undefined {
+  return logDirState.value;
+}
