+++ packages/filid/src/lib/is-debug.ts
@@ -1,0 +1,3 @@
+export function isDebug(): boolean {
+  return process.env['FILID_DEBUG'] === '1';
+}
