+++ packages/filid/src/lib/logger-types.ts
@@ -1,0 +1,5 @@
+export interface Logger {
+  debug(msg: string, ...args: unknown[]): void;
+  warn(msg: string, ...args: unknown[]): void;
+  error(msg: string, ...args: unknown[]): void;
+}
