+++ packages/filid/src/lib/reset-logger.ts
@@ -1,0 +1,6 @@
+import { logDirState } from './log-dir-state.js';
+
+/** Reset logger state (for testing). */
+export function resetLogger(): void {
+  logDirState.value = undefined;
+}
