+++ packages/filid/src/mcp/tools/debt-manage/utils/get-debt-dir.ts
@@ -1,0 +1,5 @@
+import { join } from 'node:path';
+
+export function getDebtDir(projectRoot: string): string {
+  return join(projectRoot, '.filid', 'debt');
+}
